pyproject.toml
[tool.poetry]
name = "python-regrex"
version = "0.1.15"
description = "string generator based on regex pattern"
authors = ["admin"]
packages = [{include = "src"}]
readme = "docs/README.md"

[tool.poetry.scripts]
regrex = 'src.app:main'

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = "23.2.1"
aiohttp = "3.8.4"
exrex = "0.11.0"
httpx = "0.26.0"
pyyaml = "6.0.1"
tqdm = "4.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
